+++ src/Queues/Backends/Mock_Backend.php
@@ -1,26 +1,38 @@
-<?php
-
+<?php declare(strict_types=1);
 
 namespace Tribe\Libs\Queues\Backends;
 
+use RuntimeException;
 use Tribe\Libs\Queues\Contracts\Backend;
 use Tribe\Libs\Queues\Message;
 
 /**
  * Class Mock_Backend
  *
- * A trivial backend for use when running tests
+ * A trivial backend for use when running tests.
  */
 class Mock_Backend implements Backend {
-	/** @var array[] */
+
+	/**
+	 * @var array<string, Message[]>
+	 */
 	private $queues = [];
+
+	/**
+	 * Allow returning the internal properties for tests.
+	 *
+	 * @return array<string, Message[]>
+	 */
+	public function queues(): array {
+		return $this->queues;
+	}
 
 	public function enqueue( string $queue_name, Message $m ) {
 		$this->queues[ $queue_name ][] = $m;
 	}
 
 	/**
-	 * @param string $queue_name
+	 * @param  string  $queue_name
 	 *
 	 * @return Message The first message in the queue. Nothing will be reserved.
 	 */
@@ -28,15 +40,22 @@ class Mock_Backend implements Backend {
 		if ( array_key_exists( $queue_name, $this->queues ) && ! empty( $this->queues[ $queue_name ] ) ) {
 			return reset( $this->queues[ $queue_name ] );
 		}
-		throw new \RuntimeException( 'No messages available to reserve.' );
+
+		throw new RuntimeException( 'No messages available to reserve.' );
 	}
 
 	public function ack( string $job_id, string $queue_name ) {
-		return; // does nothing
+		$id = (int) $job_id;
+
+		if ( ! isset( $this->queues[ $queue_name ][ $id ] ) ) {
+			return;
+		}
+
+		unset( $this->queues[ $queue_name ][ $id ] );
 	}
 
 	public function nack( string $job_id, string $queue_name ) {
-		return; // does nothing
+		// does nothing
 	}
 
 	public function get_type(): string {
@@ -59,4 +78,5 @@ class Mock_Backend implements Backend {
 	public function cleanup() {
 		$this->queues = [];
 	}
+
 }