+++ src/Queues/CLI/Process.php
@@ -1,22 +1,24 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Tribe\Libs\Queues\CLI;
 
 use DI;
 use Exception;
-use Throwable;
 use Tribe\Libs\CLI\Command;
-use Tribe\Libs\Queues\Contracts\Task;
+use Tribe\Libs\Queues\Message;
 use Tribe\Libs\Queues\Queue_Collection;
 use WP_CLI;
 
+use function WP_CLI\Utils\check_proc_available;
 use function WP_CLI\Utils\get_flag_value;
 
+/**
+ * Check the Queue for jobs to run.
+ */
 class Process extends Command {
 
-	public const ARG_QUEUE     = 'queue';
-	public const ARG_TIMELIMIT = 'timelimit';
-
+	public const ARG_QUEUE         = 'queue';
+	public const OPTION_TIMELIMIT  = 'timelimit';
 	public const DEFAULT_TIMELIMIT = 300;
 
 	/**
@@ -36,31 +38,40 @@ class Process extends Command {
 	 */
 	private $timelimit;
 
+	/**
+	 * Whether this server supports sub processes.
+	 *
+	 * @var bool
+	 */
+	private $proc_enabled;
+
 	public function __construct( Queue_Collection $queue_collection, DI\FactoryInterface $container ) {
-		$this->queues    = $queue_collection;
-		$this->container = $container;
+		$this->queues       = $queue_collection;
+		$this->container    = $container;
+		$this->proc_enabled = check_proc_available( 'runcommand', true );
+
 		parent::__construct();
 	}
 
-	public function command() {
+	public function command(): string {
 		return 'queues process';
 	}
 
-	public function description() {
+	public function description(): string {
 		return __( 'Process the queue for the provided queue name.', 'tribe' );
 	}
 
-	public function arguments() {
+	public function arguments(): array {
 		return [
 			[
-				'type'        => 'positional',
+				'type'        => self::ARGUMENT,
 				'name'        => self::ARG_QUEUE,
 				'optional'    => false,
 				'description' => __( 'The name of the Queue.', 'tribe' ),
 			],
 			[
-				'type'        => 'assoc',
-				'name'        => self::ARG_TIMELIMIT,
+				'type'        => self::OPTION,
+				'name'        => self::OPTION_TIMELIMIT,
 				'optional'    => true,
 				'description' => __( 'The max process execution in seconds.', 'tribe' ),
 				'default'     => self::DEFAULT_TIMELIMIT,
@@ -68,13 +79,22 @@ class Process extends Command {
 		];
 	}
 
+	/**
+	 * Execute the command.
+	 *
+	 * @param array $args
+	 * @param array $assoc_args
+	 *
+	 * @throws WP_CLI\ExitException
+	 * @return void
+	 */
 	public function run_command( $args, $assoc_args ) {
 		if ( ! isset( $args[0] ) ) {
 			WP_CLI::error( __( 'You must specify which queue you wish to process.', 'tribe' ) );
 		}
 
 		[ $queue_name ]  = $args;
-		$this->timelimit = (int) get_flag_value( $assoc_args, self::ARG_TIMELIMIT, self::DEFAULT_TIMELIMIT );
+		$this->timelimit = (int) get_flag_value( $assoc_args, self::OPTION_TIMELIMIT, self::DEFAULT_TIMELIMIT );
 
 		if ( ! array_key_exists( $queue_name, $this->queues->queues() ) ) {
 			WP_CLI::error( __( "That queue name doesn't appear to be valid.", 'tribe' ) );
@@ -88,10 +108,24 @@ class Process extends Command {
 
 		$endtime = time() + $this->timelimit;
 
+		$runtime_message = sprintf( __( 'The queue will run for %d seconds', 'tribe' ), $this->timelimit );
+
+		if ( $this->timelimit === 0 ) {
+			$runtime_message = __( 'Timelimit set to 0. The queue will run indefinitely or until memory is exhausted', 'tribe' );
+		}
+
+		WP_CLI::log(
+			sprintf(
+				__( 'Processing queue: %s. %s. Run with the --debug option to see additional details.', 'tribe' ),
+				$queue_name,
+				$runtime_message
+			)
+		);
+
 		// Run forever.
 		while ( $this->timelimit === 0 || time() < $endtime ) {
 
-			// If the queue is empty, sleep on it and then clear it again.
+			// If the queue is empty, sleep on it and then check it again.
 			if ( ! $queue->count() ) {
 				WP_CLI::debug( __( 'Queue is empty. Sleeping...', 'tribe' ) );
 				sleep( 1 );
@@ -108,24 +142,53 @@ class Process extends Command {
 
 			$task_class = $job->get_task_handler();
 
-			try {
-				/** @var Task $task */
-				$task = $this->container->make( $task_class );
-			} catch ( Throwable $e ) {
-				WP_CLI::debug( sprintf( __( 'Unable to create task instance for class "%s". Error: %s', 'tribe' ), $task_class, $e->getMessage() ) );
-				$queue->nack( $job->get_job_id() );
+			if ( ! $this->proc_enabled ) {
+				WP_CLI::warning(
+					__(
+						'The PHP functions `proc_open()` and/or `proc_close()` are disabled. Please check your PHP ini directive `disable_functions` or suhosin settings.',
+						'tribe'
+					)
+				);
 
-				return;
-			}
+				WP_CLI::debug( __( 'Running task via WP_CLI::run_command() instead, which runs in the same process.', 'tribe' ) );
 
-			if ( $task->handle( $job->get_args() ) ) {
-				// Acknowledge.
-				WP_CLI::debug( sprintf( 'ACK: %s (%s)', $job->get_job_id(), $task_class ) );
-				$queue->ack( $job->get_job_id() );
+				WP_CLI::run_command( array_merge( [
+					's1',
+					'queues',
+					'run',
+				], $job->get_args() ), [
+					Run::OPTION_JOB_ID => $job->get_job_id(),
+					Run::OPTION_QUEUE  => $queue_name,
+					Run::OPTION_TASK   => $task_class,
+				] );
 			} else {
-				WP_CLI::debug( sprintf( 'NACK: %s (%s)', $job->get_job_id(), $task_class ) );
-				$queue->nack( $job->get_job_id() );
+				$options   = $this->build_command_options( $job->get_job_id(), $queue_name, $task_class );
+				$arguments = $this->build_command_arguments( $job );
+
+				// Run in a child process, resetting state
+				WP_CLI::runcommand( sprintf( 's1 queues run %s %s', $options, $arguments ) );
 			}
 		}
 	}
+
+	protected function build_command_options( string $job_id, string $queue_name, string $task_class ): string {
+		$args = [
+			Run::OPTION_JOB_ID => $job_id,
+			Run::OPTION_QUEUE  => $queue_name,
+			Run::OPTION_TASK   => $task_class,
+		];
+
+		$options = '';
+
+		foreach ( $args as $option => $value ) {
+			$options .= sprintf( '--%s=%s ', $option, escapeshellarg( $value ) );
+		}
+
+		return $options;
+	}
+
+	protected function build_command_arguments( Message $job ): string {
+		return implode( ' ', $job->get_args() );
+	}
+
 }