+++ src/Queues/Queues_Definer.php
@@ -1,22 +1,28 @@
-<?php
-declare( strict_types=1 );
+<?php declare(strict_types=1);
 
 namespace Tribe\Libs\Queues;
 
 use DI;
+use Psr\Container\ContainerInterface;
 use Tribe\Libs\CLI\CLI_Definer;
+use Tribe\Libs\Container\Container;
 use Tribe\Libs\Container\Definer_Interface;
+use Tribe\Libs\Container\MutableContainer;
 use Tribe\Libs\Queues\Backends\WP_Cache;
 use Tribe\Libs\Queues\CLI\Add_Tasks;
 use Tribe\Libs\Queues\CLI\Cleanup;
 use Tribe\Libs\Queues\CLI\List_Queues;
 use Tribe\Libs\Queues\CLI\Process;
+use Tribe\Libs\Queues\CLI\Run;
 use Tribe\Libs\Queues\Contracts\Backend;
 use Tribe\Libs\Queues\Contracts\Queue;
 
 class Queues_Definer implements Definer_Interface {
 	public function define(): array {
 		return [
+			MutableContainer::class => static function ( ContainerInterface $c ) {
+				return ( new Container() )->wrap( $c );
+			},
 			Backend::class          => DI\create( WP_Cache::class ),
 			Queue::class            => DI\autowire( DefaultQueue::class ),
 			Queue_Collection::class => DI\create()
@@ -30,6 +36,7 @@ class Queues_Definer implements Definer_Interface {
 				DI\get( Add_Tasks::class ),
 				DI\get( Cleanup::class ),
 				DI\get( Process::class ),
+				DI\get( Run::class ),
 			] ),
 		];
 	}